refactor(common): tests.py 의 이름을 test.py로 변경
